Refactor: rename model.ts to utils.ts
