+++ pages/story/[id].tsx
@@ -28,7 +28,7 @@ import generateSocial from '../../lib/social'
 import StoryDetail from '../../components/stories/StoryDetail'
 import FloatingRibbon, { Button } from '../../components/common/FloatingRibbon'
 import HeadTags from '../../components/common/HeadTags'
-import { storyImage } from '../../components/stories/model'
+import { storyImage } from '../../components/stories/utils'
 import { GetStaticPropsResult } from 'next'
 import ErrorPage from '../404'
 import { CSSProperties } from 'react'
